Add database setup
Not making verify script fail if formatting is not successful because
it fails on some generated files, even though they're in the filter
in build gradle files. ktlintcheck task correctly ignores them
but not the format task. So if there are formatting issues that ktlint
is not able to fix on its own, the pipeline would fail and would need
to manually fix those changes.