Fix category deserialization, other minor changes
Started solving for when import subscription isn't found and found out
that categories in the get podcast by url or search API can be null.
So accounting for that.

When podcast by url isn't found, then it returns a bad request, so
capturing that error in case want to surface that to UI