Fix podcast has new episodes issue (#66)
When I added this change, podcast has new episodes boolean was always
getting written to true because we would always get a result from the
API as we get episodes since (last episode's fetch time - 1 hour).

So that last episode will always be fetched. This means that it was
resetting the has new episodes to true always.

Fix was to rely on the rows that get written to the db for that value.

Tried things where Sqlite allows getting the row number for row that
gets inserted and it's 0 if not inserted but that was yielding wrong
results for episode additional data table. Had to add a new query to
see if episode is already available in the additional data table and
then skip adding it, plus return the insert result as false even though
we still insert the episode data in the table in that case.

That is because episode data could change, but additional data is our
data and only we would change it via separate methods.