Fix work manager initialization (#100)
Noticed in debug builds at least, `EpisodeFetchWorker` was having
trouble getting initialized because `ProcessLifecycleOwner` needed
by `ForegroundStateObserver`, requires the callback to be set on
main thread. But the way koin initializes dependencies, is lazy
and so initialization can happen when the dependency is actually
needed.

Tried to use `= get()` in the worker instead of previous `by inject()`
which instantiates the object right away and not lazily but that was
throwing the same error about the callback needing to be in main
thread.

The fix was to use our own work manager factory provided via Koin
and then also let the `EpisodeFetcher` get initialized in
`MainApplication` in debug builds as well. Then the individual classes
skip doing work if we're on a debug build.