Fix episode not getting marked as completed
When auto play is not on, the currently playing episode was not getting
marked as completed, as in the function that plays the next episode,
I was exiting early if auto play was off.

This moves the mark current episode as completed to be the first line
which is always called irrespective of if auto play next is on or not.