Fix background download of episode file
When downloading new episodes in background (from the worker),
episodes are marked as need to be downloaded and not downloaded
right away because of an issue where foreground service (for
download) can not be started from background.
Instead they get downloaded if the app is brought to foreground
or when requesting download from settings.