Fix has new episodes bug
I was clearing podcast has new episodes if no new episodes were fetched
but that reset the indicator if there were new episodes previously
even if user hadn't acknowledged.

Setting the boolean to true if the episode list is not empty (new).
Setting it to false only when user acknowledges it (like before).