Fix playback not starting after service destroy
Noticed when tied player controller lifecycle to player screen, that
playback won't start again after going into the app and hitting the
play button.

Turns out that from that change, service was getting destroyed sooner,
without app being destroyed (which is expected) but in its onDestroy,
I was canceling the coroutine scope which the whole app uses. Since
that app itself was still alive, new scope wasn't created until app
restart but this was causing no coroutines to be able to run in that
same scope anymore.

Created a new coroutine scope for the media session service that doesn't
affect rest of the app