Fix queue rearragement animation (#169)
Was able to verify that queue rearrangement needing to be saved to db is
what was causing the laggy animating. Moved the episodes to an in memory
list and then tried to rearrange them with that, but added 100ms delay
with saving the list and was able to reproduce it. 

Fix was to have a queue to write changes from to database but then
reflect changes to UI right from in memory list.