Fix auto play next from background (#174)
Auto play next in queue while keeping app state as the single source
of truth, setting next episode on the player when current episode
ends doesn't work on newer versions of Android because media can't
be played from background as it starts a foreground service.a

The fix was to use media3 player's playlist to keep a copy of the
queue in there as well so that player can automatically go to the next
episode even from background.

There are several ways a queue that needs to be set on the player can
be changed
- Sleep timer end of episode
- Auto play on/off
- App queue is updated

For now, all of these actions are manually invoking `player.updateQueue`
because listening to changes in all of those values was causing issues
with playback, adding a lot of stutters. Probably need to debounce
requests to update player queue when that happens. Will explore that
option later and want to get this out right now so that the strategy
can be tested more.