Fix history screen crash
Noticed that the key was somehow still not unique on the history list
screen which not sure how that would happen since one episode could
appear only once in a single day and the key is being generated using
the date of playback for the episode.

Anyway, added session id to the key as well to see if this fixes it.