Fix back button not closing player (#216)
When player is open on a different screen than home, pressing back
was navigating out of that screen rather than closing player first.

Looks like the order of where back handler is planted matters.
Moving it to top of all screens (since the parent of all screens is
a Box) fixes it.