Fix queue rearrangement getting broken some times
This has been an ongoing issue where the queue rearrangement doesn't
work at times. Trying to drag an item just glitches and it's
impossible to place the episode item in the desired queue position.

Finally was able to figure out what was going on. It was possible
for multiple episodes to get the same queue position. That's because
on adding an episode to queue, the queue position is derived from
the max existing queue position and adding 1 to it. There's a setting
to automatically add new episodes to queue and that happens when
refreshing podcasts. However, all the refreshes happen in parallel
and so if multiple podcasts have auto add to queue on, the new
episodes would get added to the queue at the same time, getting
the same queue position. Which causes the issue

Making it so that add to queue requests are sequential now