Fix HomeScreen not loaded if started via deeplink (#272)
Fixed a weird issue where if app is launched from deeplink for
episode details and then home tab button is tapped, home screen
view state is not loaded. Gemini suggested saving and restoring
state when navigating but that didn't work.

That gave me the idea to maybe pop the home screen first if we're
navigating to home screen top level destination which fixes the
issue